==> notebooks/01_data_fetch.ipynb <==
import requests
import json
import os

#Creates data folder if it doesn't exist
os.makedirs('../data', exist_ok=True)

API_URL = "https://gis.charlottenc.gov/arcgis/rest/services/ODP/CMPD_Homicide/MapServer/0/query?where=1%3D1&outFields=*&outSR=4326&f=json"
response = requests.get(API_URL)
data = response.json()

with open('../data/homicide_raw.json', 'w') as f:
    json.dump(data, f)
