==> notebooks/02_data_cleaning.ipynb <==
import pandas as pd
import json
import os

#Load raw homicide data from local JSON file
with open('../data/homicide_raw.json') as f:
    data = json.load(f)

#Extract relevant fields from each incident
rows = []
for feature in data['features']:
    attr = feature['attributes']
    geom = feature.get('geometry', {})

    #Construct a row dictionary for each incident
    row = {
    'Latitude': geom.get('y'),
    'Longitude': geom.get('x'),
    'Incident_Date': attr.get('DATE_REPORTED'),
    'NPA': attr.get('NPA'),
    'Weapon': attr.get('WEAPON'),
    'Age': attr.get('AGE'),
    'Gender': attr.get('GENDER'),
    'Race': attr.get('RACE_ETHNICITY')
    }
    rows.append(row)
#Convert list of dictionaries to a DataFrame
df = pd.DataFrame(rows)

df.dropna(subset=['Latitude', 'Longitude'], inplace=True)
df['Incident_Date'] = pd.to_datetime(df['Incident_Date'], unit='ms')

df['Year'] = df['Incident_Date'].dt.year
df['Month'] = df['Incident_Date'].dt.month
df['Weekday'] = df['Incident_Date'].dt.dayofweek
df['Hour'] = df['Incident_Date'].dt.hour
#Filters to just Charlotte
df = df[
    (df['Latitude'] >= 34.5) & (df['Latitude'] <= 36) &
    (df['Longitude'] >= -82) & (df['Longitude'] <= -78)
]


os.makedirs('../data', exist_ok=True)
df.to_csv('../data/homicide_cleaned.csv', index=False)

df.head()
df.info()
df[['Weapon', 'Gender', 'Race']].value_counts().head()
print('File saved:', os.path.exists('../Charlotte Homicide Project/homicide_cleaned.csv'))
