==> notebooks/03_heatmap.ipynb <==
import pandas as pd
import folium
from folium.plugins import HeatMap, MarkerCluster
import os


df = pd.read_csv('../data/homicide_cleaned.csv', parse_dates=['Incident_Date'])
df.head()

df = df.dropna(subset=['Latitude', 'Longitude'])
#Creates a map based off the average lat and long
m = folium.Map(
    location=[df['Latitude'].mean(), df['Longitude'].mean()],
    zoom_start=11,
    tiles='OpenStreetMap'
)
heat_data = df[['Latitude', 'Longitude']].values.tolist()
HeatMap(heat_data).add_to(m)
os.makedirs('../outputs', exist_ok=True)
m.save('../outputs/homicide_heatmap.html')

#Layers heat data over the base map
HeatMap(
    heat_data,
    radius=25,
    blur=15,
    min_opacity=0.45,
    max_zoom=16
).add_to(m)

# marker_cluster = MarkerCluster().add_to(m)

# for _, row in df.iterrows():
#     folium.Marker(
#         location=[row['Latitude'], row['Longitude']],
#         popup=(
#             f"<b>Date:</b> {row['Incident_Date'].date()}<br>"
#             f"<b>Race:</b> {row['Race']}<br>"
#             f"<b>Gender:</b> {row['Gender']}<br>"
#             f"<b>Age:</b> {row['Age']}<br>"
#             f"<b>Weapon:</b> {row['Weapon']}<br>"
#         )
#     ).add_to(marker_cluster)


m
