==> notebooks/04_subsets.ipynb <==
import pandas as pd
import folium
from folium.plugins import HeatMap
import os


df = pd.read_csv('../data/homicide_cleaned.csv', parse_dates=['Incident_Date'])
df = df.dropna(subset=['Latitude', 'Longitude'])

def plot_heatmap(df_subset, filename):
    heat_data = df_subset[['Latitude', 'Longitude']].dropna().values.tolist()
    m = folium.Map(
        location=[df_subset['Latitude'].mean(), df_subset['Longitude'].mean()],
        zoom_start=11,
        tiles='OpenStreetMap'
    )
    HeatMap(
        heat_data,
        radius=25,
        blur=15,
        min_opacity=0.45,
        max_zoom=16
    ).add_to(m)
    os.makedirs('../outputs', exist_ok=True)
    map_path = f'../outputs/{filename}.html'
    m.save(map_path)
    print(f"✅ Map saved: ../outputs/{filename}.html")

    


plot_heatmap(df[df['Year'] == 2023], 'heatmap_2023')

plot_heatmap(df[df['Weapon'] == 'Handgun'], 'heatmap_handgun')

plot_heatmap(df[df['NPA'] == 12], 'heatmap_npa12')

plot_heatmap(df[(df['Weapon'] == 'Knife') & (df['Year'] == 2022)], 'heat_map_knife_2022')
