==> notebooks/06_ml.ipynb <==
import pandas as pd
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report
import joblib
import numpy as np

#Load the cleaned data
df = pd.read_csv('../data/homicide_cleaned.csv', parse_dates=['Incident_Date'])
df = df.dropna(subset=['Latitude','Longitude', 'Weapon'])

#Extract time features for prediction
df['Year'] = df['Incident_Date'].dt.year
df['Month'] = df['Incident_Date'].dt.month
df['Weekday'] = df['Incident_Date'].dt.dayofweek

#Prepare Features, X, and Labels, y
X = df[['Year', 'Month', 'Weekday']]
y = df['Weapon']

#Encode the Weapon Labels
encoder = LabelEncoder()
y_encoded = encoder.fit_transform(y)
X_train, X_test, y_train, y_test = train_test_split(X, y_encoded, test_size=0.2, random_state=42)

#Train a Random Forest Classifier
clf = RandomForestClassifier(random_state=42)
clf.fit(X_train, y_train)

#Evaluates the model
y_pred = clf.predict(X_test)
unique_labels = np.unique(y_test)
unique_class_names = encoder.inverse_transform(unique_labels)

all_labels = np.arange(len(encoder.classes_))
all_class_names = encoder.inverse_transform(all_labels)
print(classification_report(y_test, y_pred, labels=all_labels, target_names=all_class_names))

#Saves the model and encoder
joblib.dump(clf, '../models/weapons_predictor.pkl')
joblib.dump(encoder, '../models/weapons_encoder.pkl')
