==> notebooks/09_model_refinement_evaluation.ipynb <==
import pandas as pd
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report
from sklearn.utils import class_weight

df = pd.read_csv('../data/homicide_cleaned.csv', parse_dates=['Incident_Date'])
df = df.dropna(subset=['Latitude', 'Longitude', 'Weapon'])

le = LabelEncoder()
df['Weapon_Label'] = le.fit_transform(df['Weapon_Simplified'])

def age_group(age):
    if pd.isnull(age):
        return'Unknown'
    elif age < 18:
        return 'Minor'
    elif 18 <= age < 30:
        return 'Young Adult'
    elif 30 <= age < 50:
        return 'Adult'
    elif 50 <= age < 70:
        return 'Senior'
    else:
        return 'Elderly'
df['Age_Group'] = df['Age'].apply(age_group)
print(df['Age_Group'].value_counts())

#Season features

def season(month):
    if month in [12, 1, 2]:
        return 'Winter'
    elif month in [3, 4, 5]:
        return 'Spring'
    elif month in [6, 7, 8]:
        return 'Summer'
    else:
        return 'Fall'
df['Season'] = df['Month'].apply(season)

print(df['Season'].value_counts())

df['IsWeekend'] = df['Weekday'].apply(lambda x: 1 if x >= 5 else 0)

df['Year'] = df['Incident_Date'].dt.year
df['Month'] = df['Incident_Date'].dt.month
df['Weekday'] = df['Incident_Date'].dt.dayofweek

features = ['Year', 'Month', 'Weekday', 'Season', 'Age_Group', 'IsWeekend']
df_encoded = pd.get_dummies(df[features], drop_first=True)

X = df_encoded
y = df['Weapon_Label']

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=42, stratify=y
)

clf = RandomForestClassifier(n_estimators=100, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)
print(classification_report(y_test, y_pred, target_names=le.classes_))

#Calculate class weights based in class imbalance
weights = class_weight.compute_class_weight(
    class_weight='balanced',
    classes=le.classes_,
    y=df['Weapon_Simplified']
)

class_weights_dict = {i: weight for i, weight in enumerate(weights)}

print('Class Weights:', class_weights_dict)

param_grid = {
    'n_estimators': [100, 200],
    'max_depth': [None, 10, 20],
    'min_samples_split': [2, 5],
}

clf = RandomForestClassifier(class_weights_dict, random_state=42)

grid = GridSearchCV(clf, param_grid, cv=3, n_jobs=-1, scoring='f1_weighted')
grid.fit(X_train, y_train)

print('Best Parameters:', grid.best_params_)

best_model = grid.best_estimator_
y_pred = best_model.predict(X_test)

print('Tuned Classification Report:')
print(classification_report(y_test, y_pred, target_names=le.classes_))
